requirements.txt
# API Gateway Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Load Testing
locust==2.20.1
